requirements.txt
streamlit
easyocr
numpy
Pillow
python-Levenshtein